pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pymysql = "^1.1.1"
cryptography = "^43.0.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-stubs = "^3.15.1"
django-stubs = "^5.1.1"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
isort = "^5.13.2"
coverage = "^7.6.7"
django-stubs = "^5.1.1"
types-pymysql = "^1.1.0.20241103"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
